refactor: 이름 변경이 필요한 부분들 수정 및 필요없는 코드 삭제
